Simplify interface for inline methods.
