+++ monoruby/src/builtins.rs
@@ -34,7 +34,7 @@ pub use monoasm::*;
 pub use monoasm_macro::*;
 use monoruby_attr::monoruby_builtin;
 use num::ToPrimitive;
-pub(crate) use object::{object_send, object_send_splat, send};
+pub(crate) use object::{object_send, send};
 pub use time::TimeInner;
 
 //
