+++ monoruby/src/builtins/kernel.rs
@@ -80,7 +80,10 @@ fn object_nil(
     bb: &mut BBContext,
     callid: CallSiteId,
     _pc: BytecodePtr,
-) {
+) -> bool {
+    if !store[callid].is_simple() {
+        return false;
+    }
     if bb.is_nil(store[callid].recv) {
         ir.inline(|gen, _| {
             monoasm! { &mut gen.jit,
@@ -104,6 +107,7 @@ fn object_nil(
         });
     }
     bb.rax2acc(ir, store[callid].dst);
+    true
 }
 
 ///
