+++ monoruby/src/globals.rs
@@ -20,8 +20,13 @@ pub use store::*;
 
 pub static WARNING: std::sync::LazyLock<AtomicU8> = std::sync::LazyLock::new(|| AtomicU8::new(0u8));
 
-pub(crate) type InlineGen =
-    dyn Fn(&mut jitgen::asmir::AsmIr, &Store, &mut jitgen::BBContext, CallSiteId, BytecodePtr);
+pub(crate) type InlineGen = dyn Fn(
+    &mut jitgen::asmir::AsmIr,
+    &Store,
+    &mut jitgen::BBContext,
+    CallSiteId,
+    BytecodePtr,
+) -> bool;
 
 #[derive(Debug, Clone, PartialEq)]
 pub(crate) struct MethodTableEntry {
@@ -43,7 +48,6 @@ impl MethodTableEntry {
 
 pub(crate) const GLOBALS_FUNCINFO: usize =
     std::mem::offset_of!(Globals, store.functions.info) + MONOVEC_PTR;
-pub(crate) const OBJECT_SEND_FUNCID: FuncId = FuncId::new(3);
 
 #[derive(Clone, Debug)]
 pub(crate) struct ExternalContext {
@@ -190,19 +194,6 @@ impl Globals {
         assert_eq!(
             FuncId::new(2),
             globals.define_builtin_func_rest(OBJECT_CLASS, "", yielder)
-        );
-        assert_eq!(
-            OBJECT_SEND_FUNCID,
-            globals.define_builtin_inline_funcs_with(
-                OBJECT_CLASS,
-                "send",
-                &["__send__"],
-                crate::builtins::send,
-                Box::new(crate::builtins::object_send),
-                0,
-                0,
-                true,
-            )
         );
         globals.random.init_with_seed(None);
         crate::builtins::init_builtins(&mut globals);
